# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BeaconAI cybersecurity API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalScans: number;
  threatsBlocked: number;
  accuracyRate: number;
  communityReports: number;
}