# lib/api-zod/src/generated/types/deepfakeInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BeaconAI cybersecurity API
 * OpenAPI spec version: 0.1.0
 */
import type { DeepfakeInputMediaType } from './deepfakeInputMediaType';

export interface DeepfakeInput {
  mediaBase64: string;
  mediaType: DeepfakeInputMediaType;
}