# lib/api-zod/src/generated/types/hotspot.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BeaconAI cybersecurity API
 * OpenAPI spec version: 0.1.0
 */

export interface Hotspot {
  city: string;
  state: string;
  count: number;
  topScamType: string;
}