# lib/api-zod/src/generated/types/indicator.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BeaconAI cybersecurity API
 * OpenAPI spec version: 0.1.0
 */
import type { IndicatorRiskLevel } from './indicatorRiskLevel';

export interface Indicator {
  label: string;
  confidence: number;
  riskLevel: IndicatorRiskLevel;
}