# lib/api-zod/src/generated/types/phishingResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BeaconAI cybersecurity API
 * OpenAPI spec version: 0.1.0
 */
import type { Indicator } from './indicator';
import type { PhishingResultRiskLevel } from './phishingResultRiskLevel';
import type { PhishingResultVerdict } from './phishingResultVerdict';

export interface PhishingResult {
  verdict: PhishingResultVerdict;
  confidence: number;
  riskLevel: PhishingResultRiskLevel;
  summary: string;
  domain: string;
  indicators: Indicator[];
  recommendations: string[];
}