# lib/api-zod/src/generated/types/scamResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BeaconAI cybersecurity API
 * OpenAPI spec version: 0.1.0
 */
import type { Indicator } from './indicator';
import type { ScamResultRiskLevel } from './scamResultRiskLevel';
import type { ScamResultVerdict } from './scamResultVerdict';

export interface ScamResult {
  verdict: ScamResultVerdict;
  confidence: number;
  riskLevel: ScamResultRiskLevel;
  summary: string;
  /** @nullable */
  category: string | null;
  indicators: Indicator[];
  recommendations: string[];
}